Finished JSONGenerator
