+++ src/resources/response.json
@@ -12,7 +12,7 @@
           "country": {
             "name": "United States"
           },
-          "state": {
+          "province": {
             "name": "Wisconsin"
           },
           "city": {
